requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Automat==20.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
azure-mgmt-compute==31.0.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.1.0
azure-storage-blob==12.20.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
cloud-init==23.4.4
colorama==0.4.6
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.1
dbus-python==1.2.18
display==1.0.0
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
ecdsa==0.19.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.1
fastjsonschema==2.16.3
filelock==3.9.0
Flask==3.0.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.44.0
frozenlist==1.3.3
fsspec==2023.10.0
gitdb==4.0.11
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.17.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==21.3.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.0.333
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.45
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
mistune==2.0.5
more-itertools==8.10.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
num2words==0.5.13
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.17.0
orjson==3.10.0
packaging==23.2
pandas==2.1.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.2.0
plotly==5.18.0
portalocker==2.8.2
protobuf==4.25.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodome==3.20.0
pycryptodomex==3.17
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.1.1
pyparted==3.11.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-environ==0.4.54
python-magic==0.4.24
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2022.10.31
requests==2.31.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
SecretStorage==3.3.1
Send2Trash==1.8.0
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sos==4.5.6
SQLAlchemy==2.0.23
ssh-import-id==5.11
starlette==0.37.2
sympy==1.12
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.65.0
transformers==4.35.0
triton==2.1.0
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ufw==0.36.1
urllib3==1.26.14
uvicorn==0.29.0
validators==0.22.0
virtualenv==20.21.0
wadllib==1.3.6
WALinuxAgent==2.2.46
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.7
yarl==1.8.2
zipp==3.15.0
zope.interface==5.4.0
pyodbc
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
